68.14.3 - layout - servo

+++ layout/style/ServoStyleSet.h
@@ -23,7 +23,7 @@
 #include "nsTArray.h"
 
 namespace mozilla {
-enum class MediaFeatureChangeReason : uint8_t;
+enum class MediaFeatureChangeReason : uint16_t;
 namespace css {
 class Rule;
 }  // namespace css
